# Save generated statistics to Postgres instead of Rails cache