--- a/app/models/concerns/team_associations.rb
+++ b/app/models/concerns/team_associations.rb
@@ -17,6 +17,7 @@ module TeamAssociations
     has_many :project_groups, dependent: :destroy
     has_many :feed_teams, dependent: :destroy
     has_many :feeds, through: :feed_teams
+    has_many :monthly_team_statistics # No "dependent: :destroy" because we want to retain statistics
 
     has_annotations
   end