--- a/config/initializers/zz_open_telemetry.rb
+++ b/config/initializers/zz_open_telemetry.rb
@@ -1,7 +1,7 @@
 require 'check_open_telemetry_config'
 require 'check_open_telemetry_test_config'
 
-# Lines immediately below set any environment config that should 
+# Lines immediately below set any environment config that should
 # be applied to all environments
 ENV['OTEL_LOG_LEVEL'] = CheckConfig.get('otel_log_level')