--- a/db/migrate/20230106220307_create_monthly_team_statistics.rb
+++ b/db/migrate/20230106220307_create_monthly_team_statistics.rb
@@ -1,0 +1,28 @@
+class CreateMonthlyTeamStatistics < ActiveRecord::Migration[5.2]
+  def change
+    create_table :monthly_team_statistics do |t|
+      t.integer :conversations
+      t.integer :average_messages_per_day
+      t.integer :unique_users
+      t.integer :returning_users
+      t.integer :valid_new_requests
+      t.integer :published_native_reports
+      t.integer :published_imported_reports
+      t.integer :requests_answered_with_report
+      t.integer :reports_sent_to_users
+      t.integer :unique_users_who_received_report
+      t.integer :median_response_time
+      t.integer :unique_newsletters_sent
+      t.integer :new_newsletter_subscriptions
+      t.integer :newsletter_cancellations
+      t.integer :current_subscribers
+
+      t.datetime :start_date
+      t.datetime :end_date
+      t.string :platform
+      t.string :language
+      t.references :team, null: false, foreign_key: true, index: true
+      t.timestamps null: false
+    end
+  end
+end