… lib/tasks/data/statistics.rake
… test/lib/tasks/lapis_error_codes_test.rb
… test/lib/tasks/statistics_test.rb
… test/models/monthly_team_statistics_test.rb
… test/models/team_test.rb